+++ src/main/java/sqlline/SqlLineCompleter.java
@@ -45,4 +45,4 @@ class SqlLineCompleter
   }
 }
 
-// End SqlLinecompleter.java
+// End SqlLineCompleter.java